rename InputWrapper to FieldWrapper
