+++ stories/index.js
@@ -6,18 +6,17 @@ import './CandidateAvatar';
 import './Checkbox';
 import './Chip';
 import './ContentPlaceholder';
+import './FieldWrapper';
 import './Footer';
 import './Header';
 import './Heading';
 import './Icon';
 import './Input';
-import './InputWrapper';
 import './Link';
 import './List';
 import './LoadingSpinner';
 import './Modal';
 import './Navigation';
-import './WidthRestrictor';
 import './Pagination';
 import './ProgressBar';
 import './RadioButton';
@@ -25,3 +24,4 @@ import './Tabs';
 import './Teaser';
 import './Text';
 import './TextArea';
+import './WidthRestrictor';
